requirements.txt
pyautogui==0.9.54
Pillow>=9.4.0
numpy>=1.20.0
requests>=2.25.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0
pync>=2.0.3 